backend/requirements.txt
imbox==0.9.8
beautifulsoup4==4.12.2
requests==2.31.0
python-dotenv==1.0.0
lxml==4.9.3
